+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## 0.0.18
 - Fix duplicate state in Domoticz log for On/Off switches
 - Fix issue with reconnect after MQTT server restart (#84)
 - Handle devices connect and removal from zigbee network